oooops put the test in the wrong test module!!
my excuse is that I am on a train going at high speed towards to
London!!
